Moving VectorViewMvt to prim/mat/meta
