+++ stan/math/prim/scal/meta/size_of.hpp
@@ -2,6 +2,7 @@
 #define STAN_MATH_PRIM_SCAL_META_SIZE_OF_HPP
 
 #include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <cstddef>
 
 namespace stan {
